+++ Changelog.md
@@ -6,6 +6,24 @@ nav_order: 3
 
 # Changelog
 
+## version 20.8 :mask:
+
+### Added:
+
++ New [Radial visualization](docs/user-interface.md#mode-switches) for all modes;
++ Option to display [level (dB) scale](docs/user-interface.md#scale) on vertical axis;
++ New [**Demo** preset](docs/user-interface.md#preset).
+
+### Changed:
+
++ Improved the background image [**Pulse**](docs/user-interface.md#image-fit) effect to look more synced regardless of music style;
++ Any image located in the song's folder can now be used as album cover when a picture is not found in the song metadata (see the [documentation](docs/user-interface.md#background) for filename precedence).
+
+### Fixed:
+
++ Audio files with uppercase extensions not recognized when audioMotion was running on a standard web server.
+
+
 ## version 20.6 :mask:
 
 ### Added: